Adds a new Stat, Average, and returns the computed average of the given scores (#2)
Adds the Stat interface and the Average struct to fulfill it.
Adds Value to Scores so we can actually do something with them.
Modifies ScoreStore (and memory store) to return the names of score categories.
Reorganizes interface types into their own files. This way, test implementations can be stored with the interface they fulfill cleanly.
Implements AddAction and GetStats internal functionality in keep() and get() respectively. This is mostly for style reasons, and to make it easier to move to a concurrent model. These are what I consider the "imperative shell" of the program.
uses the empty interface (interface{}) in multiple places to allow for extension of scores and stats that may require different data types. The current implementation uses float64 for averages.